+++ frontend/src/redux/reducers/invitationsReducer.js
@@ -1,0 +1,193 @@
+const fakeCreator = {
+    name: 'John Doe',
+    profilePictureLink: 'https://www.business2community.com/wp-content/uploads/2017/08/blank-profile-picture-973460_640.png',
+};
+
+const fakeLocation = {
+    "address_components":[
+       {
+          "long_name":"Vancouver",
+          "short_name":"Vancouver",
+          "types":[
+             "locality",
+             "political"
+          ]
+       },
+       {
+          "long_name":"UBC",
+          "short_name":"UBC",
+          "types":[
+             "neighborhood",
+             "political"
+          ]
+       },
+       {
+          "long_name":"Greater Vancouver A",
+          "short_name":"Greater Vancouver A",
+          "types":[
+             "administrative_area_level_3",
+             "political"
+          ]
+       },
+       {
+          "long_name":"Metro Vancouver",
+          "short_name":"Metro Vancouver",
+          "types":[
+             "administrative_area_level_2",
+             "political"
+          ]
+       },
+       {
+          "long_name":"British Columbia",
+          "short_name":"BC",
+          "types":[
+             "administrative_area_level_1",
+             "political"
+          ]
+       },
+       {
+          "long_name":"Canada",
+          "short_name":"CA",
+          "types":[
+             "country",
+             "political"
+          ]
+       },
+       {
+          "long_name":"V6T 1Z4",
+          "short_name":"V6T 1Z4",
+          "types":[
+             "postal_code"
+          ]
+       }
+    ],
+    "adr_address":"<span class=\"locality\">Vancouver</span>, <span class=\"region\">BC</span> <span class=\"postal-code\">V6T 1Z4</span>, <span class=\"country-name\">Canada</span>",
+    "formatted_address":"Vancouver, BC V6T 1Z4, Canada",
+    "geometry":{
+       "location":{
+          "lat":49.26060520000001,
+          "lng":-123.2459939
+       }
+    },
+    "icon":"https://maps.gstatic.com/mapfiles/place_api/icons/v1/png_71/school-71.png",
+    "name":"The University of British Columbia",
+    "place_id":"ChIJAx7UL8xyhlQR86Iqc-fUncc",
+    "types":[
+       "university",
+       "point_of_interest",
+       "establishment"
+    ],
+    "url":"https://maps.google.com/?cid=14383886875425940211",
+    "html_attributions":[
+       
+    ],
+    "lat":49.26060520000001,
+    "lng":-123.2459939
+};
+
+const fakeTimezone = {
+    "value": "America/Dawson",
+    "label": "(GMT-7:00) Dawson, Yukon",
+    "offset": -7
+};
+
+const fakeSchedule = {
+    "schedule": [
+        "2022-06-22T16:00:00.000Z",
+        "2022-06-22T17:00:00.000Z",
+        "2022-06-22T18:00:00.000Z",
+        "2022-06-22T19:00:00.000Z",
+        "2022-06-22T20:00:00.000Z",
+        "2022-06-22T21:00:00.000Z",
+        "2022-06-22T22:00:00.000Z",
+        "2022-06-22T23:00:00.000Z",
+        "2022-06-23T00:00:00.000Z",
+        "2022-06-23T15:00:00.000Z",
+        "2022-06-23T16:00:00.000Z",
+        "2022-06-23T17:00:00.000Z",
+        "2022-06-23T18:00:00.000Z",
+        "2022-06-23T19:00:00.000Z",
+        "2022-06-23T20:00:00.000Z",
+        "2022-06-23T21:00:00.000Z",
+        "2022-06-23T22:00:00.000Z",
+        "2022-06-23T23:00:00.000Z",
+        "2022-06-24T00:00:00.000Z",
+        "2022-06-24T15:00:00.000Z",
+        "2022-06-24T16:00:00.000Z",
+        "2022-06-24T17:00:00.000Z",
+        "2022-06-24T18:00:00.000Z",
+        "2022-06-26T15:00:00.000Z",
+        "2022-06-26T16:00:00.000Z",
+        "2022-06-26T17:00:00.000Z",
+        "2022-06-26T18:00:00.000Z",
+        "2022-06-26T19:00:00.000Z",
+        "2022-06-26T20:00:00.000Z",
+        "2022-06-26T21:00:00.000Z",
+        "2022-06-26T22:00:00.000Z",
+        "2022-06-26T23:00:00.000Z",
+        "2022-06-27T00:00:00.000Z",
+        "2022-06-27T15:00:00.000Z",
+        "2022-06-27T16:00:00.000Z",
+        "2022-06-27T17:00:00.000Z",
+        "2022-06-27T18:00:00.000Z",
+        "2022-06-27T19:00:00.000Z",
+        "2022-06-27T20:00:00.000Z",
+        "2022-06-27T21:00:00.000Z",
+        "2022-06-27T22:00:00.000Z",
+        "2022-06-27T23:00:00.000Z",
+        "2022-06-28T00:00:00.000Z",
+        "2022-06-28T15:00:00.000Z",
+        "2022-06-28T16:00:00.000Z",
+        "2022-06-28T17:00:00.000Z",
+        "2022-06-28T18:00:00.000Z",
+        "2022-06-28T19:00:00.000Z",
+        "2022-06-28T20:00:00.000Z",
+        "2022-06-28T21:00:00.000Z",
+        "2022-06-28T22:00:00.000Z",
+        "2022-06-28T23:00:00.000Z",
+        "2022-06-29T00:00:00.000Z"
+    ],
+    "timezone": "America/Vancouver",
+    "startDate": "2022-06-23T02:59:43.289Z",
+    "selectionScheme": "linear",
+    "numDaysInput": 7,
+    "numDays": 7,
+    "hourlyChunkInput": 1,
+    "hourlyChunk": 1,
+    "timeInterval": [
+        8,
+        18
+    ]
+};
+
+
+const fakeMeetup = {
+    title: 'Party 1',
+    description: 'a description of party 1...',
+    schedule: [],
+    location: fakeLocation,
+    attendees: []
+};
+
+const initialValue = {
+    creator: fakeCreator,
+    meetup: fakeMeetup,
+    attendees: [],
+
+};
+
+
+const invitationsReducer = (state = initialValue, action) => {
+    switch (action.type) {
+        case 'SET_INVITATIONS':
+            return {
+                state
+            };
+        default:
+            return state
+    }
+}
+
+
+
+export default invitationsReducer;
